server/requirements.txt
flask
redis
flask_restx
shortuuid
openai
python-dotenv